move filter component to components folder + update imports in searchcontrol
