+++ src/components/search/SearchControl.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import Cookies from "js-cookie";
 import SearchBar from "./SearchBar";
-import FilterButton from "./Filter";
+import FilterButton from "../Filter";
 import { Affiliation, Category, IClub, School } from "@/lib/models/Club";
 import Trie from "./Trie";
 import FollowFilter from "./FollowFilter";
